remodeled pool folder and refs
