+++ carcara/src/ast/mod.rs
@@ -7,7 +7,7 @@ mod macros;
 mod context;
 mod iter;
 mod polyeq;
-mod pool;
+pub mod pool;
 pub(crate) mod printer;
 mod rc;
 mod substitution;
@@ -17,9 +17,7 @@ mod tests;
 pub use context::{Context, ContextStack};
 pub use iter::ProofIter;
 pub use polyeq::{alpha_equiv, polyeq, tracing_polyeq};
-pub use pool::TPool;
-pub use pool::TermPool;
-pub use pool::{AdvancedPools, PrimitivePool};
+pub use pool::{PrimitivePool, TPool, TermPool};
 pub use printer::print_proof;
 pub use rc::Rc;
 pub use substitution::{Substitution, SubstitutionError};
